Add support for path attributes
It seems cumbersome to have to use `from` if I wanted to use `step` or `build`
because its unclear what `from` means in this context, so instead lets add
`path` as an alias of sorts to the bare string. path and from/to are mutually
exclusive because it seems like it would be very confusing to allow `path/to`
while `path/from` definitely makes no sense.

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>