from/to uses copy not move
We may be uploading an artifact that we can read but not write (e.g. the
directory it's in is owned by root because it was produced by a
container with the Docker daemon running as root), which case this
atttempted move will fail:

    Moving [target/example-0.x.jar] to [example.jar]...
    mv: cannot move 'target/example-0.x.jar' to 'example.jar': Permission denied
    ~/builds/test-stack-i-01234567890abcdef-1/org/example

And, something running after this plugin may reasonably assume that the
file still exists at its original location after being uploaded as an
artifact. That's not highly likely, since we're doing this in a
post-command hook, but another post-command or other hook could expect
that.

The main downside of `cp` instead of `mv` is the duplication of the
data, which may be large. Generally artifact uploads won't be huge
enough for copying them to matter much, and the working directory will
be cleaned by the next build for that pipeline.